product-api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pyjwt==2.9.0
pytest==8.3.0
httpx==0.27.0
